# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "GitHubFeatured.API",
    "version": "1.0"
  },
  "paths": {
    "/Repo": {
      "post": {
        "tags": [
          "Repo"
        ],
        "summary": "Refresh the most starred repositories of each language.",
        "description": "Clear all saved repositories then search for the most starred repositories of each language and save into db.",
        "operationId": "RefreshLanguageRepositories",
        "responses": {
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Repo"
        ],
        "summary": "Returns starred repositories of each language that are saved in db.",
        "description": "Returns starred repositories of each language that are saved in db.",
        "operationId": "GetAll",
        "parameters": [
          {
            "name": "Language",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "PageNumber",
            "in": "query",
            "schema": {
              "maximum": 2147483647,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RepoModelPagedResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepoModelPagedResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepoModelPagedResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/Repo/{id}": {
      "get": {
        "tags": [
          "Repo"
        ],
        "summary": "Return a specific starred repository that's saved in db.",
        "description": "Return a specific starred repository that's saved in db.",
        "operationId": "GetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RepoModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepoModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepoModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "data": {
            "nullable": true
          },
          "succeeded": {
            "type": "boolean"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RepoModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "gitHubId": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "private": {
            "type": "boolean"
          },
          "ownerUser": {
            "type": "string",
            "nullable": true
          },
          "htmlUrl": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "fork": {
            "type": "boolean"
          },
          "url": {
            "type": "string",
            "nullable": true
          },
          "homepage": {
            "type": "string",
            "nullable": true
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "language": {
            "type": "string",
            "nullable": true
          },
          "visibility": {
            "type": "string",
            "nullable": true
          },
          "score": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "pushedAt": {
            "type": "string",
            "format": "date-time"
          },
          "stargazersCount": {
            "type": "integer",
            "format": "int64"
          },
          "watchersCount": {
            "type": "integer",
            "format": "int64"
          },
          "forksCount": {
            "type": "integer",
            "format": "int64"
          },
          "hasIssues": {
            "type": "boolean"
          },
          "openIssuesCount": {
            "type": "integer",
            "format": "int32"
          },
          "hasProjects": {
            "type": "boolean"
          },
          "hasDownloads": {
            "type": "boolean"
          },
          "hasWiki": {
            "type": "boolean"
          },
          "hasPages": {
            "type": "boolean"
          },
          "hasDiscussions": {
            "type": "boolean"
          },
          "archived": {
            "type": "boolean"
          },
          "disabled": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "RepoModelPagedResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RepoModel"
            },
            "nullable": true
          },
          "succeeded": {
            "type": "boolean"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      }
    }
  }
}